prevent unintended behaviour and crash in protostar
protostar stores the played hand in poker_hand to be accessed later, in after. if a poker hand is not allowed, joker_main is not called but after is. so after is using an outdated or empty poker_hand value in this case. this fixes that by adding guards to prevent after triggering when the poker_hand is empty and emptying it when after triggers so the value cannot be used later